Move all widgets in "widgets" folder
