--- libs/search-widget/src/Widget.scss
@@ -1,1 +1,0 @@
-@import "common/common-style";
