+++ libs/search-widget/src/widgets/video-widget/SearchResults.scss
@@ -1,4 +1,4 @@
-@import "../common/common-style";
+@import "../../common/common-style";
 
 .error {
   align-items: center;
