+++ libs/search-widget/src/widgets/widget/Widget.scss
@@ -1,0 +1,1 @@
+@import "../../common/common-style";
