+++ libs/search-widget/src/widgets/widget/index.ts
