+++ libs/search-widget/src/widgets/widget/lib.ts
